add wishlist to component did mount and refactor arrow and carousel styling
